added support to decode test dev/eval1

+++ egs/material/s5/local/preprocess_test.sh
@@ -15,38 +15,50 @@ mkdir -p $datadev
 
 # 1. create the reference transcript $datadev/reftext
 
-ls -d $audio_path/transcription/* > "list.tmp"
-rm -rf {zero,brac,all}.tmp
-rm -rf all.tmp.sort
+if [ $(basename $datadev) == 'analysis1' ]; then
+  ls -d ${audio_path_analysis1}/transcription/* > "list.tmp"
+  rm -rf {zero,brac,all}.tmp
+  rm -rf all.tmp.sort
 
-while read line; do
-  h=$(head -1 "$line")
-  if [[ $h == "0"* ]]
-  then                                                                     
-    # starts with 0.000"
-    echo $line | while read ref; do
-      s=${ref##*/}
-      awk '{l[NR] = $0} END {for (i=1; i<=NR-1; i++) print l[i]}' "$ref" | \
-        cut -f1,2 --complement | tr '\n' ' ' | \
-        awk '{$0="'${s%.transcription.txt}' " $0}1' >> zero.tmp;
-    done
-  else
-    # starts with [0.000]"
-    echo $line | while read ref; do
-      s=${ref##*/}
-      awk 'NR%2==0' "$ref" | tr '\n' ' ' | \
-        awk '{$0="'${s%.transcription.txt}' " $0}1'>> brac.tmp;
-    done
-  fi
-done < "list.tmp"
+  while read line; do
+    h=$(head -1 "$line")
+    if [[ $h == "0"* ]]
+    then                                                                     
+      # starts with 0.000"
+      echo $line | while read ref; do
+        s=${ref##*/}
+        awk '{l[NR] = $0} END {for (i=1; i<=NR-1; i++) print l[i]}' "$ref" | \
+          cut -f1,2 --complement | tr '\n' ' ' | \
+          awk '{$0="'${s%.transcription.txt}' " $0}1' >> zero.tmp;
+      done
+    else
+      # starts with [0.000]"
+      echo $line | while read ref; do
+        s=${ref##*/}
+        awk 'NR%2==0' "$ref" | tr '\n' ' ' | \
+          awk '{$0="'${s%.transcription.txt}' " $0}1'>> brac.tmp;
+      done
+    fi
+  done < "list.tmp"
 
-cat zero.tmp brac.tmp > all.tmp
-mv all.tmp $datadev/reftext
-rm -rf {zero,brac,list}.tmp
+  cat zero.tmp brac.tmp > all.tmp
+  mv all.tmp $datadev/reftext
+  rm -rf {zero,brac,list}.tmp
+fi
 
 # 2. create wav.scp, utt2spk, spk2utt files
 
-find $audio_path/src -name "*.wav" \
+wav_path=
+if [ $(basename $datadev) == 'analysis1' ]; then
+  wav_path=${audio_path_analysis1}/src
+elif [ $(basename $datadev) == 'test_dev' ]; then
+  wav_path=${audio_path_dev}/src
+elif [ $(basename $datadev) == 'eval1' ]; then
+  wav_path=${audio_path_eval1}/src
+fi
+[ -z ${wav_path} ] && echo "$0: test data should be either analysis1, test_dev, or eval1." && exit 1
+
+find ${wav_path} -name "*.wav" \
   | while read file; do id=$(basename $file | awk '{gsub(".wav","");print}'); \
   echo "$id sox $file -r 8000 -b 16 -c 1 -t wav - |"; done > \
   $datadev/wav.scp
